Added more plots/tools for VBF closure, MET studies and trigger studies

+++ MyAnalysis/util/MyJet.h
@@ -23,7 +23,7 @@ class MyJet: public TLorentzVector {
         };
         virtual ~MyJet();
 
-        void SetJVT(int j) {
+        void SetJVT(double j) {
             jvt = j;
         };
